Split draw generators by method used
This commit splits the power-pairing and random draw generators into
subclasses for the method used for pairing, the default being "australs"
for power-pairing and "swap" for random. Then, the new "graph" method
could be implemented without as close control flow.

The specific/overridden methods for specific methods are now mixins,
which are then combined with the base classes for full functionality.